refactor: divided services between favorites and notes
