+++ src/services/notesService.ts
@@ -1,6 +1,7 @@
 import * as notesRepository from "@/repositories/notesRepository";
 import * as favoritesRepository from "@/repositories/favoritesRepository";
 import { INote } from "@/protocols";
+import { findFavoritesIds, rewriteConsultDatabase } from "./favoritesService";
 
 export async function insertNewNote(newNote: INote) {
   const note = await notesRepository.insertNote(newNote);
@@ -20,14 +21,6 @@ export async function findOtherNotes() {
   return rewriteConsultDatabase(otherNotes);
 }
 
-export async function findFavoriteNotes() {
-  const idsFavorites = await findFavoritesIds();
-
-  const favoriteNotes = await favoritesRepository.findAllFavoriteNotes(idsFavorites);
-
-  return rewriteConsultDatabase(favoriteNotes);
-}
-
 export async function removeNote(noteId) {
   const removedNote = await notesRepository.deleteNote(noteId);
 
@@ -40,33 +33,7 @@ export async function editColorNote(noteId, color: string) {
   return updatedNote;
 }
 
-export async function insertFavoriteNote(noteId) {
-  const favoritedNote = await favoritesRepository.insertFavoriteNote(noteId.id);
-  return favoritedNote;
-}
-
-export async function deleteFavoriteNote(noteId) {
-  const deletedFavorite = await favoritesRepository.deleteFavoriteNote(noteId);
-  return deletedFavorite;
-}
-
 export async function updateNote(note) {
   const updatedNote = await notesRepository.updateNote(note);
   return updatedNote;
-}
-
-//utils
-async function rewriteConsultDatabase(notes) {
-  const rewriteConsult = notes.map((note) => {
-    const { _id, ...rest } = note;
-    return { id: _id, ...rest };
-  });
-
-  return rewriteConsult;
-}
-
-async function findFavoritesIds() {
-  const favorites = await favoritesRepository.findFavoriteIds();
-  const favoriteIds = favorites.map((favorite) => favorite.note_id);
-  return favoriteIds;
 }
